+++ extensions/asana/src/hooks/useMyTasks.ts
@@ -5,9 +5,10 @@ import { handleUseCachedPromiseError } from "../helpers/errors";
 
 const { showCompletedTasks } = getPreferenceValues<Preferences.MyTasks>();
 
-export function useMyTasks(workspace?: string) {
-  return useCachedPromise((workspace) => getMyTasks(workspace, showCompletedTasks), [workspace], {
+export function useMyTasks(workspace?: string, query?: string) {
+  return useCachedPromise((workspace, query) => getMyTasks(workspace, showCompletedTasks, query), [workspace, query], {
     execute: !!workspace,
+    keepPreviousData: true,
     onError(error) {
       handleUseCachedPromiseError(error);
     },